Don't use batching on GenerateServiceMetadataSources Output (#111)
Batching on the Outputs of GenerateServiceMetadataSources results in many, many calls to this target. The target itself is working in bulk and isn't doing anything that needs this output batching, so all that's happening is that more time is spent in the build performing needless target calls.

Changing this to return the Item list directly results in one call, with one pass through the `ReferencePathWithRefAssemblies` items.