Remove AspNetCore dependency in Npgsql EF Core component (#427)
The Npgsql EF Core component currently adds the EF developer page in development environments. This means the Npgsql EF Core component is dependent on the ASP.NET Core shared framework, which brings more dependencies on it in other app models (ex. Worker apps).

Removing this functionality because this dependency is not desirable and we don't have it in the SQL Server component.

Fix #146