Default Azure SDK tracing bool to the true if the environment variable or runtime config is set (#1084)
* Default Azure SDK tracing bool to the true if the environment variable or runtime config is set

With the recent update to the new Azure SDK, the tracing is no longer experimental and can be enabled without the environment variable / AppContext switch. So enable Tracing by default for all Azure components except for ServiceBus. ServiceBus tracing is still experimental. So default the Tracing value based on whether the switch is enabled. This allows users to just set 1 setting - the Azure switch - to turn on tracing for ServiceBus.

Fix #742