Reduce CPU while polling for console log changes (#1354)
The current implementation polls the file system for changes to stdout/stderr files on disk. This polling causes high CPU while logs are visible in the UI.

To reduce the problem, we introduce a small sleep that happens only when we are caught up with the contents of the file. The delay of 100ms is forever in computer time (i.e. it's a low duty cycle activity) while also being very fast for a human's perception.

Note that if logs start streaming in fast, we will not hit this sleep. It only occurs when we've read everything. If in the subsequent 100ms new logs arrive, they'll all be processed in the next iteration of the loop.

Longer term we hope to have a console log streaming API from DCP for this.