Move console log API from ResourceViewModel to IResourceService (#1391)
`ResourceViewModel` will soon be split into `ResourceSnapshot` and `ResourceViewModel`, where the snapshot will be immutable and the view model will be mutable for the UI to modify.

The view would use the `ResourceViewModel.LogSource` property to populate the `ConsoleLogs` component.

As we won't be able to send that property over the wire, we need to create a service that will achieve this.

To do this, we remove the `LogSource` property and create `IResourceService.SubscribeConsoleLogs` that provides the same interface.

We also replace the `ILogSource` interface, with its four members, with `IAsyncEnumerable<T>`, which gives a simpler implementation and enables language features like `await foreach`.

Another change in support of this was adding stdout and stderr file paths to `ExecutableViewModel`. These will be removed if DCP exposes a service for streaming logs.

The replacement of `ILogSource` required some restructuring of how we obtained console logs. The invocation patterns differed a fair bit, but the code ended up being cleaner. The diff is challenging though. It may be easier to read each commit separately.

---

* Spelling

* Seal class

Prevents warnings in dispose methods about suppressing finalization.

* Formatting

* Update regex description and options

- Add `ExplicitCapture` to avoid capture, unwrapping three instances of `(?:...)` from the pattern. Though I don't think even this is needed as we only call `IsMatch` so it doesn't have to allocate a `Match` result object.
- Updated documentation on the pattern, which had fallen out of date.
- Unified `x1B` and `\u001b` in regex pattern, as they're the same value.
- Add `CultureInvariant` as everything here is ASCII/Latin1.

* Move StringComparers into Utils namespace

* Remove LogParserOptions

* LogViewer changes

- Add class doc comment.
- Remove queue. Instead, have calls to `SetLogSourceAsync` wait on a new `TaskCompletionSource _whenDomReady`.
- Remove `_renderComplete` flag. This is now tracked by `_whenDomReady`.
- Allow stderr and stdout to be interleaved in the stream.
- Use a new `CancellationSeries` to ensures new data sources cancel prior ones.

* Add CancellationSeries

* Rename IResourceService.Subscribe to SubscribeResources

We will soon have another subscribe method, so need to differentiate by type or purpose.

* Don't do work for empty batches

Avoids an empty list allocation, an enumerator, and a call through the JS layer.

* Make properties required

Also reorder, to separate immutable and mutable properties.

* Add StdOutFile and StdErrFile to ExecutableViewModel

We'll use these for logging shortly.

* Make ProcessSpec immutable

Well, other than `EnvironmentVariables` being a mutable collection.

* Add extension method: Channel<T>.GetBatches

Reads chunks, grabbing all available messages and returning them as one. This can give a better downstream experience, as there's less per-item overhead.

* Add ResourcePublisher.TryGetResource

Allows lookup of a resource by name.

* Specify Channel<T> options for perf

Having only a single reader and writer simplifies coordination of the queue.

Allowing synchronous continuations avoids the overhead of task scheduling by having the writer execute the reader's action directly.

* Remove ILogSource

* API docs

* Merge handling of stdout and stderr

* Add IResourceService.SubscribeConsoleLogs

- Create a `ConsoleLogPublisher` to go alongside `ResourcePublisher`.
- Converts `FileLogSource` and `DockerContainerLogService` changed to implement `IAsyncEnumerable<IReadOnlyList<(string Content, bool IsErrorMessage)>>`. It's similar in spirit to `ILogSource`, but is simpler to use and has direct language support (`yield return` and `await foreach`).
- Both log sources unified handling of stdout and stderr, to reduce the amount of code.
- File log now:
  - Uses a regex to split logs read from a file.
  - Preserves blank lines. Extra logic was added to avoid adding extra blank lines between each "chunk", as files end with EOL characters. We have to ignore the last line if it's empty.
- Docker log now:
  - Uses one channel instead of two
  -

* Reorder members and convert property to field

* Extend test

* Review feedback