Specify assembly name for eShopLite's app host
The app host exposes this via `Microsoft.Extensions.Hosting.IHostEnvironment.ApplicationName`, which is documented as returning the assembly name. Any `.AppHost` suffix is removed before the name returned.

Because the name of this project is juse `AppHost`, the dashboard for this project would only show `AppHost`, which is misleading.

This sets the assembly name to have the expected naming convention, without requiring the project name to change and look different to sibling projects in eShopLite.