Ensure the page title displays
With the introduction of gRPC between dashboard and app host, the dashboard can render its UI before the application information request returns.

One negative consequence of this is that the UI's page title and header area do not show the application name. When this data arrives later asynchronously, the UI is not updated.

This change introduces a way for the UI to refresh at a later time, once the application name is available.