Fix bug in FileUtil.FindFullPathFromPath
This method attempts to find a given executable on the user's path. It does this by appending a platform-specific executable suffix (e.g. `.exe` on Windows), however it does this twice -- once before the loop and once within. This means it ends up checking for files such as `c:\folder\in\my\path\docker.exe.exe` (with double extension) which is unlikely to be found.

The fact that this has not caused any problem before now (on Windows at least) suggests to me that perhaps this method is not needed.