Validate tokens can be cancelled (#1658)
These components use the cancellation token to clean up the resources they provide. We expect all callers to provide a token that will be cancelled at some point.

This change adds validation that the tokens are actually cancelable, protecting against the case that a caller indavertently passes `CancellationToken.None` or `default(CancellationToken)`.