Improve environment variable parsing
During debugging I noticed a scenario where the returned environment variable was empty, which led to an exception that `""` was an invalid URI. This change ensures that empty or whitespace strings are treated the same as nulls, such that the default is used.

Additionally, support for ignoring of empty URLs is included too. This fixes cases like `http://foo;` (trailing semicolon).