Fix ObjectDisposedException in CancellationSeries (#1677)
When a token is requested, the prior one will be cancelled and its CTS disposed. If two concurrent requests are made quickly enough, it's important that the disposal doesn't trigger an `ObjectDisposedException`.

To address that, we request theCTS's `Token` property _before_ there's any chance for the CTS to be disposed.