Use consistent comparer for resource names (#1710)
In #1691 the semantics around comparisons of resource names changed to make them case-insensitive. This change needs to be applied in some other places, for consistency.

We have a `StringComparers` class which denotes the comparison style for different types of values. This change extends the use of that type, and changes the value for `ResourceName` from `Ordinal` to `OrdinalIgnoreCase`.

It also moves that internal type into the `Aspire` namespace so that it's picked up by both projects that link it (`Aspire.Hosting` and `Aspire.Dashboard`) without needing to choose a namespace from either one, which then wouldn't make sense in the other.