Rename 'samples' to 'integrations' (#1689)
* Rename 'samples' to 'integrations'

To avoid potential confusion, we rename this folder. These are not good samples for users to refer to as an example of how to build .NET Aspire applications. There's non-standard stuff in these projects to support development within this repo.

Also add a README.md file in that folder to direct people to the sample repo.

* Rename "integrations" to "playground"