Add public API tracking
This adds the `Microsoft.CodeAnalysis.PublicApiAnalyzers` package to the dashboard and hosting projects, though it could be added more broadly if deemed helpful.

As we approach v1, we should be very intentional about what APIs are expected to be public. I have the feeling that several public APIs we have today don't need to be public and that we may regret making them public in future.