Fix presentation of resource exit code
The resource exit code is an integer, but gRPC `Value` cannot represent integers (only `Number`, akin to `double`). Therefore we convert the integer to a string and send that.

The previous conversion used format string `N` which produces values of form `1.00`. This change moves to `D`, which produces values of form `1`. Negative exit codes are supported too.