Default exit code to null rather than -1
Exit codes only exist once the process actually exits. While a process is running, we need a value to represent the absence of an exit code.

Previously this was `Conventions.UnknownExitCode` which had value -1.

With this change we make the value nullable, and use null to indicate no value is available. This avoids the potential for confusion between integer values, and forces the consumer to think about and cater for the absence of a value.