Map DCP's unknown container exit code to null in resource service
DCP uses a sentinel value of -1 to indicate that a container's exit code is unknown. In our resource service protocol we use a null for this value. We cannot use -1, as that is a valid exit code for Windows proceses.

This change maps a container exit code of -1 to a null value, for container resources only. This avoids us displaying -1 in the dashboard UI, which is confusing.