Fix merging code coverage results
The second stage processes the *cobertura.xml files to generate a
report. And for that it collects these xml files from all the jobs.
Earlier commits in this PR did not have unique names across jobs, like
for linux, and windows. Which meant that the report generation step only
handled one set of the xml files.

Fix that by using the original unique filenames for the merged xml
generated from the helix, and non-helix tests.