[tests] Add support for running tests on helix (#1917)
* [tests] Add support for running tests on helix

With this PR, tests will be run in two steps on CI:

1. On the build machine - which is the same as they are run currently.
   This will be run in a single step, from the source repo with built
   `artifacts`.

2. On helix - for this the tests are archived to a `.zip` file, and then
   sent to helix with any dependencies.

Additionally, tests can be marked to always be run on helix or the build
machine. Other tests can be run on either.

- a new property `$(DefaultToRunningTestsOnHelix)` when true would run
  all the tests that are not build-machine-only to run on helix.
  - when the property is !true, all the tests that can run build-machine
    will run there. This is the default.

This adds new properties that affect CI builds:

- `$(IsTestSupportProject)` - which is for a project that should be run
  as a test project itself.
- `$(IsHelixOnlyTestProject)` - such a test can only be run on helix.
- `$(IsNonHelixOnlyTestProject)` - such a project must always run on the
  build machine, IOW, with the source repo.

- Also, a default `.runsettings` file is used to allow easier setting of
  loggers, and other configuration for the tests.

* [tests] Add the CI side of support for running tests on helix

All the tests will still continue to use dotnet-coverage. But since
tests can run on more than one machine now, there is an additional CI
step that merges all the coverage xml files.

* send-to-helix.proj: extract dotnet-coverage version from dotnet-tools.json

* CI: Increase timeout from 30->45mins, to account for helix waiting on agents

* Address review feedback from @ joperezr

* Address review feedback

.. and use a single property named `$(RunTestsOnHelix)` that defaults to
`false`.

* Address review feedback, and modify the original template for send-to-helix.yml

* Minor PR feedback

* Instead of modifying the template from arcade, use a copy in pipelines/templates

* send-to-helix.yml - minor cleanup

* Fix merging code coverage results

The second stage processes the *cobertura.xml files to generate a
report. And for that it collects these xml files from all the jobs.
Earlier commits in this PR did not have unique names across jobs, like
for linux, and windows. Which meant that the report generation step only
handled one set of the xml files.

Fix that by using the original unique filenames for the merged xml
generated from the helix, and non-helix tests.

---------

Co-authored-by: Eric Erhardt <eric.erhardt@microsoft.com>