Update protocol for resource commands
As we've started investigating implementing these commands, the protocol proved insufficient.

This adds the resource name to the request, and clarifies that the name on the command is a display name for the command, rather than the name of the resource. `ResourceCommandRequest` no longer does double-duty as both definition of the command and request to execute one. We split the former out into `ResourceCommand`.