Fix ServiceBus endpoint logic (#2385)
In previous versions of the ServiceBus SDK, it didn't accept `https://` URIs as the fully qualified namespace, but the ResourceManager APIs only return URIs with `https://`. So we had logic in our bicep template to remove the https. However, this doesn't work with the runtime component because it expected sb://.

The fix is to update to the latest ServiceBus version, which accepts https:// for the namespace URI. And also remove our bicp logic to just pass the endpoint through unmodified.

Also clean up the Directory.Packages.props file.

Fix #2378