Azure provisioning in the dashboard (#2552)
This change makes it possible (and cleaner IMO) to publish and log data about a resource to the dashboard and does that to show errors when creating resources and for showing provisioning progress when using the Aspire.Hosting.Azure.Provisioning package.  This change consolidates the previously made annotations for publishing and subscribing to both resource updates and logs general purpose services instead of annotations.