Enable DATAS GC for the Dashboard
In .NET 8, the .NET runtime team developed a new GC mode named "DATAS (Dynamic Adaptation To Application Sizes)". This is like Server GC, but it doesn't try to consume all the memory the machine is given. It tries to dynamically adapt to how much the app needs. You can read more about this in
https://maoni0.medium.com/dynamically-adapting-to-application-sizes-2d72fcb6f1ea.

Enable this GC mode for the Dashboard in order to consume less memory.