[tests] Restructure helix project, and targets to work better
Currently, `tests/send-to-helix.proj` builds once and constructs various
`@(HelixWorkItem)`s for end-to-end tests, and basic tests that don't
need a workload.

Because of the way Helix SDK works, this ends up sharing some stuff like
`HelixPreCommand`, and `HelixPostCommand`. But the different kinds of
tests don't need the same pre/post commands, and the targets need to be
careful not to step on each other's items/properties.

To avoid all this, and keep it clean a new top-level project
`tests/send-to-helix-ci.proj` is added which essentially builds the
existing project multiple times, once for each test type (called Test
Category here).

Thus you have:

```
send-to-helix-ci.proj runs the following in parallel:
    => send-to-helix-inner.proj testCategory=basictests
        -> send, and wait for the helix jobs
    => send-to-helix-inner.proj testCategory=endtoendtests
        -> send, and wait for the helix jobs
```