Resource service supports API key authentication (#3400) (#3582)
* Make CompareHelpers shared

* Resource service supports API keys

Unless unsecured, the app host will generate an API key and pass it to the dashboard via an environment variable. The dashboard then includes this key in a header for all gRPC calls. The app host's resource service validates that the expected key is received and rejects requests where the key is omitted.

* Remove test's exposure to ambient environment variables

* Review feedback

* Require authenticated user

* Renaming