Fix Keyed Redis
When adding 2 Redis server instances and try to use "keyed" services to talk to each, only the first connection is used for both. The 2nd connection doesn't get used.

This is because we are using a singleton `OptionsFactory<ConfigurationOptions>` that only knows about the first StackExchangeRedisSettings registered, and always uses its ConnectionString.

The fix is to teach the factory about all the Settings objects being registered, and have it find the correct one given the "options name" when it is being created.

Fix #3873