Add option to specify context service in DbContext
New code was added to allow the specification of a context service (interface) when registering a DbContext. This provides the ability to resolve the context from a container when using different services such as Azure SQL, MS SQL Server, Oracle Database and Cosmos DB. Corresponding unit tests were penned to validate the functionality across the respective databases.