Initial prototype for #3052
Since the CDK doesn't currently support configuring extensions on the PG Server resource, we have to generate the Bicep on the fly to do that. There's no easy hook that can be used after the original call to PublishAsAzurePostgresFlexibleServer to have it output values from the bicep, so instead we have to provide our own configureResource method.