Allow mutating Azure construct resources (#4434)
* Allow mutating Azure construct resources

This commit adds a new method `ConfigureConstruct` to the `AzureConstructResourceExtensions` class. The method changing any existing azure construct resource.

The changes include modifying the `AzureConstructResource` class to have an internal setter for the `ConfigureConstruct` property. Additionally, a new overload of the `ConfigureConstruct` method is added to the `IResourceBuilder<T>` interface.

Fixes #4424